+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This file contains "important" commits, small ones will probably not be added.
 
+### 200227 - 1 [2.0.25]
+- New: Git support, see modified files right from the explorer
+
 ### 200226 - 1 [2.0.24]
 - New: Notification construcotr